fix #2262 - VesselTarget.Hooks (and Parts separated)
