+++ src/kOS/kOS.csproj
@@ -168,6 +168,8 @@
     <Compile Include="Suffixed\VectorRenderer.cs" />
     <Compile Include="Suffixed\VesselAlt.cs" />
     <Compile Include="Suffixed\VesselEta.cs" />
+    <Compile Include="Suffixed\VesselTarget.Hooks.cs" />
+    <Compile Include="Suffixed\VesselTarget.Parts.cs" />
     <Compile Include="Suffixed\WaypointValue.cs" />
     <Compile Include="Execution\TransferManager.cs" />
     <Compile Include="Suffixed\Widget\Slider.cs" />
